实现用户分页查询，优化修改用户信息

+++ src/views/organization/user/index.vue
@@ -3,24 +3,38 @@
     <div class="filter-container">
 
       <!--查询条件-->
-      <el-input @keyup.enter.native="handleFilter" style="width: 200px;" class="filter-item"
-                :placeholder="$t('user.username')" v-model="listQuery.username">
+      <el-input @keyup.enter.native="handleFilter"
+                style="width: 200px;"
+                class="filter-item"
+                :placeholder="$t('user.username')"
+                v-model="listQuery.username">
       </el-input>
-      <el-input @keyup.enter.native="handleFilter" style="width: 200px;" class="filter-item"
-                :placeholder="$t('user.mobile')" v-model="listQuery.mobile">
+      <el-input @keyup.enter.native="handleFilter"
+                style="width: 200px;"
+                class="filter-item"
+                :placeholder="$t('user.mobile')"
+                v-model="listQuery.mobile">
       </el-input>
-      <el-select clearable style="width: 90px" class="filter-item" v-model="listQuery.status" :placeholder="$t('table.status')">
-        <el-option v-for="item in userStatus" :key="item" :label="$t('user.status.'+item)" :value="item"></el-option>
+      <el-select clearable
+                 style="width: 90px"
+                 class="filter-item"
+                 v-model="listQuery.status"
+                 :placeholder="$t('table.status')">
+        <el-option v-for="item in userStatus"
+                   :key="item"
+                   :label="$t('user.status.'+item)"
+                   :value="item">
+        </el-option>
       </el-select>
 
       <!--动作按钮-->
       <el-button class="filter-item" type="primary" v-waves icon="el-icon-search" @click="handleFilter">
         {{$t('table.search')}}
       </el-button>
-      <el-button class="filter-item" style="margin-left: 10px;" @click="handleCreate" type="primary" icon="el-icon-edit">
+      <el-button class="filter-item" type="primary" v-waves icon="el-icon-edit" @click="handleCreate">
         {{$t('search.add')}}
       </el-button>
-      <el-button class="filter-item" type="primary" :loading="downloadLoading" v-waves icon="el-icon-download"
+      <el-button class="filter-item" type="primary" v-waves icon="el-icon-download" :loading="downloadLoading"
                  @click="handleDownload">
         {{$t('search.export')}}
       </el-button>
@@ -53,9 +67,9 @@
       </el-table-column>
 
       <!--<el-table-column class-name="status-col" :label="$t('table.status')" width="80">-->
-        <!--<template slot-scope="scope">-->
-          <!--<el-tag :type="scope.row.status | statusFilter">{{$t('user.status.'+scope.row.status)}}</el-tag>-->
-        <!--</template>-->
+      <!--<template slot-scope="scope">-->
+      <!--<el-tag :type="scope.row.status | statusFilter">{{$t('user.status.'+scope.row.status)}}</el-tag>-->
+      <!--</template>-->
       <!--</el-table-column>-->
 
       <el-table-column width="160px" align="center" :label="$t('table.updatedTime')">
@@ -82,12 +96,12 @@
         </template>
       </el-table-column>
 
-      <el-table-column align="center" :label="$t('table.action')" width="150">
+      <el-table-column align="center" :label="$t('table.action')" width="200">
         <template slot-scope="scope">
           <el-button type="primary" size="mini" @click="handleUpdate(scope.row)">
             {{$t('table.edit')}}
           </el-button>
-          <el-button type="danger"  size="mini" @click="deleteUser(scope.row.id)">
+          <el-button type="danger" size="mini" @click="deleteData(scope.row.id)">
             {{$t('table.delete')}}
           </el-button>
         </template>
@@ -96,16 +110,21 @@
 
     <!--翻页工具条-->
     <div class="pagination-container">
-      <el-pagination background @size-change="handleSizeChange" @current-change="handleCurrentChange"
+      <el-pagination background
+                     @size-change="handleSizeChange"
+                     @current-change="handleCurrentChange"
                      :current-page="listQuery.current"
-                     :page-sizes="[10,20,30, 50]" :page-size="listQuery.size"
+                     :page-sizes="[10, 20, 30, 50]"
+                     :page-size="listQuery.size"
                      layout="total, sizes, prev, pager, next, jumper" :total="total">
       </el-pagination>
     </div>
 
     <!--添加或编辑对话框-->
     <el-dialog :title="$t('table.' + dialogStatus)" :visible.sync="dialogFormVisible">
-      <el-form :rules="rules" ref="dataForm" :model="temp" label-position="left" label-width="90px" style='width: 80%; margin-left:30px;'>
+      <el-form :rules="rules" ref="dataForm" :model="temp" label-position="left"
+               label-width="120px"
+               style='width: 80%; margin-left:60px;'>
         <el-form-item :label="$t('user.name')" prop="name">
           <el-input v-model="temp.name" placeholder="Please input a name"></el-input>
         </el-form-item>
@@ -116,10 +135,8 @@
           <el-input v-model="temp.mobile" placeholder="Please input mobile number"></el-input>
         </el-form-item>
         <el-form-item :label="$t('user.password')" prop="password">
-          <el-input v-model="temp.password" placeholder="Please input password"></el-input>
-        </el-form-item>
-        <el-form-item :label="$t('user.confirmPassword')" prop="confirmPassword">
-          <el-input v-model="temp.confirmPassword" placeholder="Please input confirmPassword"></el-input>
+          <el-input :disabled="dialogStatus=='edit'"
+                    v-model="temp.password" placeholder="Please input password" show-password></el-input>
         </el-form-item>
         <el-form-item :label="$t('user.description')" prop="description">
           <el-input :rows="3"
@@ -133,7 +150,7 @@
       <div slot="footer" class="dialog-footer">
         <el-button @click="dialogFormVisible = false">{{$t('table.cancel')}}</el-button>
         <el-button v-if="dialogStatus=='create'" type="primary" @click="createData">{{$t('table.confirm')}}</el-button>
-        <el-button v-else type="primary" @click="updateData">{{$t('table.confirm')}}</el-button>
+        <el-button v-if="dialogStatus=='edit'" type="primary" @click="updateData">{{$t('table.confirm')}}</el-button>
       </div>
     </el-dialog>
   </div>
@@ -145,6 +162,7 @@
 
   export default {
     name: 'userManagement',
+    // 水波文效果
     directives: {
       waves
     },
@@ -153,25 +171,27 @@
         list: null,
         total: 0,
         listLoading: true,
+        // 查询参数
         listQuery: {
           status: 'ok',
           current: 1,
           size: 10
         },
+        // 用户状态
         userStatus: ['lock', 'deleted', 'ok'],
         dialogStatus: 'create',
         dialogFormVisible: false,
+        // 表单校验规则
         rules: {
           username: [{ required: true, message: 'username is required', trigger: 'blur' }],
           name: [{ required: false, message: 'name is required', trigger: 'blur' }],
           mobile: [{ required: true, message: 'mobile is required', trigger: 'blur' }],
-          password: [{ required: true, message: 'password is required', trigger: 'blur' }],
-          confirmPassword: [{ required: true, message: 'confirmPassword is required', trigger: 'blur' }]
+          password: [{ required: true, message: 'password is required', trigger: 'blur' }]
         },
+        // 创建或修改用户临时对象
         temp: {
           username: '',
           name: '',
-          confirmPassword: '',
           password: '',
           description: '',
           mobile: ''
@@ -180,6 +200,7 @@
       }
     },
     filters: {
+      // 用户状态标签样式
       statusFilter(status) {
         const statusMap = {
           lock: 'danger',
@@ -189,10 +210,14 @@
         return statusMap[status]
       }
     },
+    // 页面加载完成后显示列表页
     created() {
       this.getList()
     },
     methods: {
+      /**
+       * 查询列表
+       */
       getList() {
         this.listLoading = true
         getList(this.listQuery).then(response => {
@@ -201,19 +226,29 @@
           this.listLoading = false
         })
       },
+      // 查询过滤器
       handleFilter() {
         this.listQuery.current = 1
         this.getList()
       },
+      /**
+       * 修改每页显示条数
+       */
       handleSizeChange(val) {
         this.listQuery.size = val
         this.getList()
       },
+      /**
+       * 跳转到指定页
+       */
       handleCurrentChange(val) {
         this.listQuery.current = val
         this.getList()
       },
 
+      /**
+       * 弹出创建用户表单
+       */
       handleCreate() {
         this.temp = {}
         this.dialogStatus = 'create'
@@ -222,6 +257,9 @@
           this.$refs['dataForm'].clearValidate()
         })
       },
+      /**
+       * 创建用户
+       */
       createData() {
         this.$refs['dataForm'].validate((valid) => {
           if (valid) {
@@ -238,7 +276,9 @@
           }
         })
       },
-
+      /**
+       * 弹出修改用户表单
+       */
       handleUpdate(row) {
         this.temp = Object.assign({}, row) // copy obj
         this.dialogStatus = 'edit'
@@ -247,6 +287,9 @@
           this.$refs['dataForm'].clearValidate()
         })
       },
+      /**
+       * 修改用户信息
+       */
       updateData() {
         this.$refs['dataForm'].validate((valid) => {
           if (valid) {
@@ -263,17 +306,36 @@
           }
         })
       },
-      deleteUser(id) {
-        deleteUser(id).then(() => {
-          this.$notify({
-            title: '删除成功',
-            message: '删除成功',
-            type: 'success',
-            duration: 2000
+      /**
+       * 删除用户
+       */
+      deleteData(id) {
+        this.$confirm('此操作将删除该用户, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          center: true,
+          type: 'warning'
+        }).then(() => {
+          deleteUser(id).then(() => {
+            this.$notify({
+              title: '删除成功',
+              message: '删除成功',
+              type: 'success',
+              duration: 2000
+            })
+            this.getList()
           })
-          this.getList()
         })
       },
+      /**
+       * 重置用户密码
+       */
+      resetPass(id) {
+
+      },
+      /**
+       * 导出列表
+       */
       handleDownload() {
         console.log('download')
       }
